fix(leaderboard): totalClicks int4 overflow + admin give-all-cards
- FIX (leaderboard accuracy): (game_state->>'totalClicks')::int overflows
  int4 at 2.1B. An idle game with auto-clickers exceeds that, and the cast
  then throws 'integer out of range' — failing the ENTIRE leaderboard query
  (the glitchy/inaccurate board). Cast to ::bigint instead (both the
  all-time and period leaderboard filters).
- admin: POST /admin/users/:discordId/give-all-cards — gives one of every
  catalog card the user doesn't own (testing convenience).

Suite green (105).

+++ src/models/score.model.js
@@ -157,7 +157,7 @@ const getTopScores = async (limit = 50, offset = 0) => {
         updated_at
       FROM scores
       WHERE score > 0
-        AND COALESCE((game_state->>'totalClicks')::int, 0) > 0
+        AND COALESCE((game_state->>'totalClicks')::bigint, 0) > 0
         AND discord_id != '000000000000000001'
       ORDER BY score DESC,
         COALESCE((game_state->>'ascensionLevel')::int, 0) DESC,
@@ -286,7 +286,7 @@ const getTopScoresByPeriod = async (period = 'all', limit = 50, offset = 0) => {
         updated_at
       FROM scores
       WHERE score > 0
-        AND COALESCE((game_state->>'totalClicks')::int, 0) > 0
+        AND COALESCE((game_state->>'totalClicks')::bigint, 0) > 0
         AND discord_id != '000000000000000001'
         ${timeFilter}
       ORDER BY score DESC,
